Refactor large CLI commands into modules  and add e2e test workflow
Command files were growing too large (up.rs: 1541 lines, agent.rs: 1013 lines, volume.rs: 898 lines, status.rs: 557 lines), making them hard to maintain and navigate

Split each command into a directory with focused submodules: types, format, http, display, etc. Each file now has a single responsibility with max ~300 lines. Also added a GitHub Actions e2e workflow that tests the full spuff lifecycle using the Docker provider
