+++ src/cli/commands/agent/mod.rs
@@ -1,0 +1,17 @@
+//! Agent commands
+//!
+//! Commands for interacting with the spuff-agent running on VMs
+//! or Docker containers.
+
+mod docker;
+mod exec;
+mod format;
+mod http;
+mod logs;
+mod status;
+mod types;
+
+// Re-export public functions
+pub use exec::exec;
+pub use logs::{activity, exec_log, logs};
+pub use status::{metrics, processes, status};
